+++ CHANGELOG.md
@@ -1,6 +1,25 @@
 # Change Log
 
 
+## Unreleased
+
+### Added
+
+- [Dig](http://go.uber.org/dig) dependency injection container
+
+### Changed
+
+- Renamed `NewApplication` to `New`
+
+### Removed
+
+- Logger constructors
+- Error handler constructors
+- Tracing support
+- Application closer
+- Arbitrary entries
+
+
 ## 0.3.0 - 2017-10-09
 
 ### Added